Fixed problem introduced into mesh exporter.
The exporter was asuming that any mesh with x vertices
would have its first x indices match the same index
as before optimization, but it is not the case,
the default Blender cube does meet this criteria and
I didn't test my other meshes until today.

Also brought back multithreading for some of the map calls.