Fixed Linux modelviewer.
The model viewer app now runs properly again,
though libOpenGLRenderer.so needs to be renamed to just
OpenGLRenderer OR game.config needs to be changed to
point to the actual so name.
Either way, the path to the SO needs to be added to
the LD_LIBRARY_PATH environment variable.