Added orange.svg image for testing.
GraphicsMagick is now a build time
dependency to generate PNG images
out of SVG files, this is temporary
while I decide whether to add it to
the runtime, brew my own svg->img code
and/or make the engine handle SVG natively.

Either way, this has to do more with keeping
binary files out of the repository than a preference
for vector graphics.