Adding source code for glslang compiler.
This is a modified copy of the StandAlone
directory from the Khronos glslang project.

In particular, os dependent threads have been replaced
with std::thread, and counters dump have been removed.
The change is meant to be able to compile the code
with the vanilla development libraries from glslang,
which for some reason do not include os dependent headers
or libraries.

More changes still to come.