Refactored Program into Pipeline
The program class is now the pipeline class
to better reflect its function and to adere
to the idea that the focus is on Vulkan and future APIs
rather than sticking with old constructs.

That said, a program is a core part of the OpenGL
pipeline which is something that does exist even if
not explicitly acnowledged by the API...
or I am making things up, either way, its better to think
in terms of pipelines rather than programs.