Added code to make use of shader properties (Uniforms)...
in Vulkan, refactored VulkanPipeline to hold descriptors,
descriptor sets, descriptor set layouts and descriptor pools.

Currently there is one pool per pipeline, this may not be
optimal in all cases, but for now, it is what works.

Rendering the simple phong cube is off,
so that's what's next in the fix menu.

I am giving up on trying to fix Linux build from Windows for now,
I'll upload a build fix once I have verified the build is
actually fixed on a Linux box.