Some Memory leak hunting.
After realizing that I can use the Crt Dbg Hook
to set a breakpoint at the allocation as it happens,
I found out the leaks that happen by just running
modelviewer and closing it come from glslang.

I filed a bug report:

https://github.com/KhronosGroup/glslang/issues/928

Since I am not that familiar with glslang code and
the code in question is a bit complicated to grasp.

Right now this isn't a show stopper for me,
so I won't be working on a fix for the time being.