Added animation file proto and Blender exporter.
The exporter works and seems to output correct data...
just like the skeleton one, we'll see once I write the C++ side.

Unlike the mesh and skeleton this format does have a lot of loss
in relationship with the old format, which was something I was proud of
though looking at it now, was incomplete.

For now all the animation optimizations I had will have to wait,
this format is basically raw animation data.