Changed Mesh exporter and Model file format.
The mesh exporter now has two modes,
a single mesh export that exports the only the active
mesh object, defaulting its name to the object's name,
or a multiple mesh export that exports all meshes
in the scene as their own separate file into a selected
folder. This should help the asset creation pipeline.

At this moment I am unsure as to whether or not
I want to keep triangle groups in a mesh,
they were originaly meant to separate parts of the mesh
that used a different material and/or pipeline,
but having multiple meshes solves the same problem and
is much less complex codewise.

The model format is in a transitioning state, originaly,
I just kept it as collection of single references,
binding a mesh to a pipeline, now that I need to have
multiple meshes per model, I realized I would need a material
per mesh if only to change its diffuse map, and then, why not
being able to use a different pipeline per mesh as well?
So I am making the required changes, they are not complete,
I do NOT want to keep the mesh{ mesh {} } construct in model files,
but I dont know what to name the set of pipeline-material-mesh.

I'll figure it out.