Implemented OpenGLMaterial class.
Mostly removed stuff from the convoluted mess
that was acquiring locations, offsets, indexes and
what not from the OpenGL 3.x days. Then I moved that
to the material class.

Changed the default material in pipeline to be a shared pointer
as thats what I need to create the one in the opengl pipeline
and eventually the vulkan one. This meant I no longer need
a default constructor or move assigment, so they are gone because

Less code is always better.

... so KISS YAGNI.

Had to bump the shader version to 430,
I really want to work on a shader editor,
but changing what works and will probably
work for the forseable future will only bring down my pace.