Started refactoring OpenGL renderer windows.
This may break the Linux build... but what doesn't?

Also realized that it may be better to use shared pointers to
back reference renderers, and windows may become first class citizen
classes IE: you actually get to manage them.

Dunno, will have to sleep over it.

All this is to silence that LOT of messages you get when closing
the model viewer on OpenGL mode.