Linux build fix.
I am considering renaming AeonGames::Window to
something else as it clashes with X11's Window handle.
as it is the global namespace provides a solution to deal
with the ambiguity but I don't like partial solutions,
either we add the namespace to everything or nothing,
and I'd like to have it on nothing.

Anyway this is a similar situation of having to use the
AeonGames namespace inside the same namespace in a previous commit.