Fixed some issues in Vulkan Renderer.
This allows the renderer to work with non-skeleton models.
There's still more refactoring comming, so it is unlikely
that I'll implement skeleton handling code in Vulkan before
I have a better codebase to work with.

I also found some potential issues with the VulkanRenderer::Render
function I want to look into,
the function, as it stands, is pretty much an aberration
on both renderers anyway.