Adjustments to the Resource Cache function.
The cache function AeonGames::Get was passing the key
as the first argument of the resource constructor,
while this is exactly what we want for generic resources
that take a std::string filename as constructor argument, is not
quite optimal when what you're passing is a smart pointer...
then again maybe it is, I was not seeing leaks before,
but I dont now either, so it may had only affected performace.

BUT anyway, now all previously nice and clean calls
take the same argument twice, and the ones that do not take
a string are taking a raw ptr which is what I though would be
pretty much unique.

I realized that, as is, the Get function is not fool proof,
a final user may call it at different times providing a different key
each time defeating its purpose, so, it may be a good idea to move
generic resource creation to a factory Function in
AeonGames::AeonEngine. That's for another time though.

Next, I'll add a proper back reference to the renderers to their
allocated resources.