Vulkan skeletal animation code.
Added initial code for model instance class since we're going to
need it moving forward.

With Vulkan, We'll need separate skeleton uniform buffers,
suck that each instance will have to have its own,
so, might as well just introduce the instance concept,
and beging bridging the gap between the scenegraph and the renderer.