Partially incorporated ModelInstance class into body of code.
Also did some cleanup of old code and added some notes on
ideas to explore later on.

The circular reference renderer->model->renderer issue doesn't seem
to be as critical as I though, but the model->renderer-resource
relationship is something that worries me.

You have to load the renderer specific resource in order to be able
to render it, in doing so the resource keeps a reference to the
generic resource, so now the generic resource may not be destroyed until
the renderer resource is destroyed, and this is done via the
UnloadModel function in the renderer which takes the generic resource
reference as parameter. It is posible to lose track of the generic
resource pointer outside the renderer, in which case the renderer
will keep the renderer resources as well as the generic resources
alive in memory until the renderer itself is deleted.

This isn't good, so I am thinking on a solution, but I do want to advance
development as well.

The Load/Unload functions are unfortunate as well, but they are a
necesary evil, since you do not want to do lazy construction during a
the game loop.