Adding initial static geometry Blender exporter code.
This is mostly a copy-paste of the exporter for the AeonEngine 2.
It DOES NOT work as it is right now, but that's ok because there
is no way to visualize the exported file anyway.

This format was supposed to be part of my master's degree thesis,
but that felt flat mostly due to documentation, I still
believe the collision algorithm IS quite solid.

There is also the start of the new protobuffer geometry format.