Removing some Vulkan validation layer errors.
For some reason the code works without validation,
but validation fails when enabled... better play by the rules.

Validation is disabled/commented out still,
but some changes have been made with it enabled,
I did not want to leave a broken codebase,
but it seems that if I want to use push constants for matrices
I'll have to limit myself to just two of them...
That or go back to an UBOs.

Note to self: investigate Ring Buffers.