Changed Renderer data loading to be done 'lazily' when needed.
Best course of action was to add a hidden render model interface
unique pointer to the model class to be used by renderer.

This way the model render data is deleted when the model itself
is deleted and there is no need to keep track of objects in the
renderer side... we'll see how long this lasts.

A cache function is still needed to preload scenes.