Added initial code for Dependency Map template.
Not much to see here, just keeping record not to lose any work done.

This class is meant to behave as a topologically sorted map,
such that if iterated, later elements depend on earlier elements.
An element must be defined by a Key and a vector of keys,
which represent the Key's dependencies.
A Key dependency does not have to exist at the time of insertion in the
map.