Node and Scene classes refactored into Tree class.
This is a transitional commit,
I wanted to templatize the Tree class,
but due to how transformations are handled,
and some other operations that also require tree navigation
at the node level, there is no simple way to
separate tree functionality from Transform, AABB and so on.

So I am creating this commit in case I ever want to go back,
and templatize the Tree class, but otherwise the rest of the
work will be split into a regular Tree.h header and a Tree.cpp
implementation.