Work on mesh rendering from Window classes.
Added most of the required code,
But have not tested it yet.

Instanced rendering is in, but topology is not
yet done, so rendering the editor grid will not
work.

I am not sure I like how I have to create a full new
pipeline to render lines instead of triangles for
Vulkan while its always been an argument to
the drawing functions in OpenGL...

Should I keep implicit pipeline instances in Vulkan
and emulate OpenGL or should I require multiple
explicit OpenGL pipelines to emulate Vulkan?

I'll probably start with the later,
but the former could be much more memory efficient
in the long run.