Work on dynamic node creation, and node properties.
Just a start, for now, the world editor can create
any node types registered with RegisterNodeConstructor,
but cannot yet modify its properties.

Virtual mutators have been added to the node class
for property aggregation, but I am not sure/not to happy
about how it needs to be implemented as it is right now,
its placing a lot of the burden on whoever wants to implement
new nodes.

Node::Update is no longer abstract.