Added package reading code from old engine.
There is a lot to refactor on this code,
and is not tested other than the fact that it compiles.
The class interface may or may not change.

While loading resources is the next development step,
loading resources from a pack file is not necesarily
a priority, I am adding this code much like the static level
geometry code, for it to serve as a basis for future work.