More Package class refactoring.
This time, things will break in Linux and MSVC,
Lucky for me MSYS2 updated g++ to 8.2,
and since protobuf 3.6.0.1 is broken on Visual Studio,
I am not in a rush to have builds pass there.

Ubuntu 16 still has gcc 5.4, not sure about 18,
but support for filesystem there is suposed to
be much better than Windows. We'll see.