Fixing old resources.
The mesh node works in the editor
but knowing the files' crc32s
is necesary to actually display the meshes.
Materials using textures are unlikely to work
at this time.

There seems to be an issue with depth buffer tests on Vulkan,
this may be an AMD only problem, and the issue may already be fixed
on newer drivers/sdks. OpenGL seems unafected.