Fixing Vulkan debug layer warnings.
Depth buffer is still wrong,
and there is an issue with the world editor grid pipeline
I am still not sure how to fix, but at least it seems like
despite the warning, the pipeline does work.
It may be an issue with the glslang compiler
since it pops up right at the create module call.