Removing Node inheritance and Node properties [EPIC].
This is an EPIC refactoring commit.

Moved Nodes plugin to Component plugin and
removed the MeshNode class files.

Controls for node transforms have been added to the World Editor.
The tree view for properties has been moved to the component group box,
and will now show the fields in the protobuf message for selected
components.
NodeModel has been heavily downgraded and now just shows a list of
components.

Also changed the pre-commit hook so it is no longer necesary to
manually run the format-code target.