Adding some idea code on how to implement an API for...
Protocol buffer messages as properties.

This code does nothing since it is not used,
the idea is to make message wrapper fields
mutable and move them to the component class.

This is not as easy due to repeated and message fields
having corner cases to handle, but this code is a start.

The idea is to have each field set and get value functions,
and make repeated fields be a single field with multiple
values rather than multiple fields with single values.