Revert "Adding some idea code on how to implement an API for..."
This path was becoming a roadblock and added more complexity
than actually needed to be. The world editor worked as expected
at 5 commits ago, and all I really needed is a way let components
know their pb message data had changed. I came up with
many ideas on how to do that, but the design just kept getting
more and more complicated, not to mention having to maintain
extra objects with redundant data just to keep track of PB fields.
This is fine for the editor but not for the runtime.

So, in the end I think the best solution is to clean up the
the MessageWrapper class (perhaps merge it with the MessageModel
instead), clean up the SceneWindow class code for message writting
(move it to the model or the wrapper), and finally, let the component
expose its PB Message as mutable but add a function to let
the component know something changed in the message.

This reverts commit 18e09cf879b73fae006209a350fe4b76e18b4370.
This reverts commit d68aa0d05a98d8d4192f2206fdbc3106e74ed1d0.
This reverts commit 2b5bb3aba258d368a696e8bbaf8e48aebb9d9d9a.
This reverts commit 0baea8bc4cfb88402c3f800ca3aa1daad656c97a.
This reverts commit 509fb5ecc1ea16b7d93d9532d4d8a1eab68867cb.