Work on texture mapping pipelines.
Textures work with the OpenGL Renderer,
the Vulkan render seems to have trouble
loading images in the middle of a renderpass,
so I will have to rework how/when
render resources are loaded.
I was not too happy with that code anyway.

Also found GTK+ to be less cooperative
than expected, so I'll be taking a different
UI/HUD approach.