Changes to make Image use RenderImage field.
Still not 100% convinced a global Renderer is the solution,
but not satisfied that IRender* interfaces are a good solution either.

Should an application be able to have multiple renderers?

R/ There is no reason why an application should not,
or rather no hard limitation as to why not.
A system window may only be mapped to one renderer at a time,
but an application may have different windows for different renderers,
this seems impractical but again, no hard limitation.
It is possible and required to be able to have no renderer,
the use case being a network game server.