Rename the NodeData class to Component
Task #3

I once thought the idea of a handful of components and a multitude of data nodes was a good one,
but I can't think of a single good use for separate component and data properties.

Component classes with no member variables have the same overhead(none)
as data filled component classes, so I removed the old component implementation...
but now the Component class is called NodeData. Change that.