New collision geometry export code.
Not sure if I will keep the GTY name,
as it is better to keep the collision info
separate from the rendering info and the space
partitioning info.

Also it doesn't seem to make much sence to allocate
a brush per polygon/face if the full mesh is already convex.