Work on Linux Standalone OpenGL Window.
There is a LOT of work here, the least of it has anything to do
with Linux Standalone OpenGL Window.

The World Editor stopped working with the OpenGL renderer
for some reason, but to be honest I am surpriced it even worked
in the first place to begin with.

The FrameBuffer Configuration used to create the GLX context
as part of Renderer initialization happened to match the visual
for the Qt Window and thus, a single context was used.
Making the context current for the window was no problem.

Until recently.

Something may have changed at the Qt library which made this lucky
coincidence work, and now I was getting a BadMatch error on
glXMakeCurrent calls. I though it was the Display, but not really,
then the visual, but hardwiring the visual to whatever visual the
Window had only showed a black screen, and there is no way to guess
what the Window visual will be at Renderer creation, so that was no
good.

In the end, I created a shared GLX context for the Qt Window,
but OpenGL does not share VAO's, so now I have a different problem,
and since I get an error when binding arrays, I don't even know if this
will be a proper solution.

So, there is two options here:

1- Finally learn how to use FBOs to render on a headless context
   and then move the result to the window context.
2- Remove use of multiple VAOs, which some say are slow anyway.

I kind of like 1 better than 2, I think this is what Vulkan does anyway,
but 2 is something I could get done in a couple of days and quickly find
out if the shared context solution works.

Anyway, for now I'll be moving work to Windows as there is nothing
broken there yet.