Use a C cast to handle divergences on VK_NULL_HANDLE definition
The Vulkan SDK used to define VK_NULL_HANDLE as 0, but latelly
is being defined as nullptr when C++ is detected at the header.
This causes problems when casting with reinterpret cast as well
as leaving the cast out, so revert to C type cast which should
handle both use cases.