Centralized resource loading code in Resource class.
A slight difference on a Skeleton::Load funcion was likely
the source of the issue seen in Linuxwhere loading animations
caused a crash, so it is better to keep all duplicate code in
a single spot.

Also, I no longer care much about public vs internal headers,
when the time comes to build a proper no core build SDK, I'll do it.