Added more detailed setup for Qt's surface format.
OpenGL renderer fails because the frame buffer config selected
by the renderer algorithm is not compatible with the one selected
by Qt, this is likely due to the fact that samples are disabled
by default on the Qt side, while the renderer selects the config
with the most samples.

Still needs to be tested on Linux and a sensible
number of samples should be set.