Vulkan Renderer compiles now, Does not work.
Renamed UsePipeline to BindPipeline which will now
just bind the pipeline and shall not set matrices, material or
skeleton, these will be set on different Renderer function calls.

This change is due to how Vulkan differs from OpenGL which has no
push constants or descriptor sets and the Window class even if
API dependent should not be poking around into Renderer resources,
IE: I could leave the Generic Renderer API as is but add a getter
for the bound pipeline to the Vulkan Renderer so the Window can
retrieve it and use it, but thats just bad design.

Also still To do: completely move CommonWindow outside the engine core.