Removing the public Window class.
Too much platform dependent code was seeping into renderers,
and keeping renderers separate from windows was just unmaintenable,
you needed a window to create a renderer and a renderer to create
a window, so instead now you do need a PLATFORM specific window to
create a renderer, and you need to take care of all the platform
specific nuances other than the ones related to the window render
surface.

After a renderer has been created with a reference to a platform
window, it is posible to attach multiple platform dependent windows
to the same renderer and share the renderer resources, but they must
be created in the same manner as the initial window.

This is a HUGE commit that includes lots of fixes and changes as it
should be for a commit that changes the very foundation of the
whole design, Window management has been left to the end user
and a many things have been simplified. The OpenGL renderer
now has proxy functions to better detect issues, though not
as robust as Vulkan layers, it does help a lot.

Finally all platform dependent code now resides at the renderers
and the demo application, where at the renderers platform dependent
code has been kept to a very minimum, and the core and component
libraries no longer have any kind of dependency on the renderer
libraries.

This change should cover both Windows and Linux builds.