Added Base64 encoding and decoding functions.
Binary strings in text protobuf files are too long,
sometimes 3 or 4 times the original binary array.

Binary files such as images are troublesome for git,
which is why their inclussion has been kept to a minimum
in favour of SVG images, but it will be necesary at some point
to add all kinds of 'source' binary files, these shall
be base64 encoded.