Added proposed new xml file for worldeditor grid pipeline
This xml file format is meant as a transitional format
to be converted into a binary to be consumed by the engine.

It is meant to be more generic than the previous Protobuffer approach
to allow for all types of shaders and not specialize on skeletal/mesh/matrix
based pipelines with limited vertex attributes or reserved unused uniforms.